+++ winui3/SquirrelNotifier.WinUI3.Tests/Helpers/UrlValidatorTests.cs
@@ -1,0 +1,46 @@
+// <copyright file="UrlValidatorTests.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+using FluentAssertions;
+using SquirrelNotifier.WinUI3.Helpers;
+using Xunit;
+
+namespace SquirrelNotifier.WinUI3.Tests.Helpers;
+
+public class UrlValidatorTests
+{
+    [Theory]
+    [InlineData("https://github.com/scottlz0310/squirrel-notifier/pull/51", true)]
+    [InlineData("https://github.com/scottlz0310-user/another-repo/pull/123?query=param", true)]
+    [InlineData("http://github.com/scottlz0310/squirrel-notifier/pull/51", false)] // http instead of https
+    [InlineData("https://evil.com/github.com/scottlz0310/squirrel-notifier", false)] // wrong host
+    [InlineData("https://github.com/scottlz0310/squirrel-notifier;rm -rf", false)] // unsafe characters
+    [InlineData("", false)]
+    [InlineData(null, false)]
+    public void IsSafeGitHubUrl_ShouldValidateCorrectly(string? url, bool expected)
+    {
+        // Act
+        bool result = UrlValidator.IsSafeGitHubUrl(url);
+
+        // Assert
+        result.Should().Be(expected);
+    }
+
+    [Theory]
+    [InlineData("https://github.com/scottlz0310/squirrel-notifier/pull/51", "scottlz0310/squirrel-notifier", 51, true)]
+    [InlineData("https://github.com/scottlz0310/squirrel-notifier/pull/51", "scottlz0310/another-repo", 51, false)] // repo mismatch
+    [InlineData("https://github.com/scottlz0310/squirrel-notifier/pull/51", "scottlz0310/squirrel-notifier", 52, false)] // pr mismatch
+    [InlineData("https://github.com/SCOTTLZ0310/SQUIRREL-NOTIFIER/pull/51", "scottlz0310/squirrel-notifier", 51, true)] // case insensitive
+    [InlineData("http://github.com/scottlz0310/squirrel-notifier/pull/51", "scottlz0310/squirrel-notifier", 51, false)] // unsafe scheme
+    [InlineData("", "repo", 1, false)]
+    [InlineData(null, "repo", 1, false)]
+    public void IsSafeGitHubUrl_WithRepoAndNumber_ShouldValidateCorrectly(string? url, string repository, int prNumber, bool expected)
+    {
+        // Act
+        bool result = UrlValidator.IsSafeGitHubUrl(url, repository, prNumber);
+
+        // Assert
+        result.Should().Be(expected);
+    }
+}