+++ winui3/SquirrelNotifier.WinUI3/Services/INotificationService.cs
@@ -1,6 +1,5 @@
-// <copyright file="INotificationService.cs" company="PlaceholderCompany">
-// Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
+using System;
+using SquirrelNotifier.WinUI3.Models;
 
 namespace SquirrelNotifier.WinUI3.Services;
 
@@ -9,4 +8,10 @@ internal interface INotificationService
     void Initialize();
 
     void NotifyReviewEventReceived(string? message, string? recommendedNextAction);
+
+    void NotifyReviewEvent(ReviewEvent reviewEvent);
+
+    event EventHandler<ReviewEvent>? ReviewEventReceived;
+
+    event EventHandler? OpenAppRequested;
 }